requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlalchemy
pydantic
python-multipart
pillow
PyMuPDF
aiofiles
request
